небольшие правки
