+++ packages/client/src/сonstants/text.ts
@@ -22,6 +22,7 @@ export const REGISTRATION_BUTTON_TEXT = 'Регистрация';
 
 // profile
 export const PROFILE_CHANGE_DATA = 'Изменить данные';
+export const CHANGE_PASSWORD_TEXT = 'Изменить данные';
 export const EDIT_CHANGE_DATA = 'Редактировать данные';
 export const AVATAR_TEXT = 'Загружаемый аватар';
